// src/app/maps/interfaces/interface.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
  type: string;
  query: string[];
  features: Feature[];
  attribution: string;
}

export interface Feature {
  id: string;
  type: string;
  place_type: string[];
  relevance: number;
  properties: Properties;
  text: string;
  place_name: string;
  matching_text?: string;
  matching_place_name?: string;
  center: number[];
  geometry: Geometry;
  context: Context[];
}

export interface Context {
  id: string;
  wikidata: Wikidata;
  text: Text;
  short_code?: ShortCode;
}

export enum ID {
  Country8707 = 'country.8707',
  Place370691 = 'place.370691',
  Region58371 = 'region.58371',
}

export enum MapboxID {
  DXJuOm1IeHBsYzo1QU0 = 'dXJuOm1ieHBsYzo1QU0',
  DXJuOm1IeHBsYzpCYWdE = 'dXJuOm1ieHBsYzpCYWdE',
  DXJuOm1IeHBsYzpJZ00 = 'dXJuOm1ieHBsYzpJZ00',
}

export enum ShortCode {
  AE = 'ae',
  AEDu = 'AE-DU',
}

export enum Text {
  Dubai = 'Dubai',
  UnitedArabEmirates = 'United Arab Emirates',
}

export enum Wikidata {
  Q612 = 'Q612',
  Q613 = 'Q613',
  Q878 = 'Q878',
}

export interface Geometry {
  coordinates: number[];
  type: string;
}

export interface Properties {
  foursquare?: string;
  landmark?: boolean;
  address?: string;
  category?: string;
  accuracy?: string;
}